pyproject.toml
[project]
name = "eink-dashboard"
version = "0.1.0"
description = "E-ink dashboard for Raspberry Pi"
authors = [
    {name = "Vince Bello",email = "vincent.bello@gmail.com"}
]
readme = "README.md"
requires-python = "^3.12"
dependencies = [
    "requests (==2.32.3)",
    "pillow (==10.4.0)",
    "tzdata (==2024.1)",
    "pytz (==2024.1)",
    "google-auth (==2.35.0)",
    "google-auth-oauthlib (==1.2.1)",
    "google-auth-httplib2 (==0.2.0)",
    "google-api-python-client (==2.151.0)",
    "gtfs-realtime-bindings (==1.0.0)",
    "protobuf (==4.25.5)",
    "fonttools (>=4.62.1,<5.0.0)",
    "brotli (>=1.2.0,<2.0.0)"
]


[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
